requirements.txt
asgiref==3.5.2
boto3==1.24.63
botocore==1.27.63
certifi==2022.6.15
charset-normalizer==2.1.1
clickhouse-driver==0.2.4
clickhouse-sqlalchemy==0.2.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
greenlet==1.1.3
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
minio==7.1.11
openapi-codec==1.3.2
psycopg2==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.40
sqlparse==0.4.2
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
